# spring-server/target/classes/swagger_GamificationPlatform.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: Gamification
  description: An API to implement gamification features in the swagger-2.0 specification. 
    
    ToDo 
      (- /rules)
    
  contact:
    name: Guillaume Serneels
    email: guillaume.serneels@heig-vd.ch
    url: www.heig-vd.ch
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
host: localhost
basePath: /api
schemes:
  - http
paths:
  /badges:
    get:
      summary: Returns every badges
      description: |
        Returns every badges
      operationId: findBadges
      produces:
        - application/json
      parameters:
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: badges response
          schema:
            type: array
            items:
              $ref: '#/definitions/badge'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    post:
      summary: Creates a new badge
      description: |
        Creates a new badge
      operationId: addBadge
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: badge
          in: body
          description: Badge object to add to the store
          required: true
          schema:
            $ref: '#/definitions/badge'
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: badge response
          schema:
            $ref: '#/definitions/badge'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /badges/{badgeId}:   
    get:
      summary: Returns a single badge
      description: |
        Returns a single badge
      operationId: findBadge
      produces:
        - application/json
      parameters:
        - name: badgeId
          in: path
          description: ID of badge to fetch
          required: true
          type: integer
          format: int64
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: badge response
          schema:
            $ref: '#/definitions/badge'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    put:
      summary: Update an existing badge
      description: |
        Update an existing badge
      operationId: UpdateBadge
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: badge
          in: body
          description: Badge object to add to the store
          required: true
          schema:
            $ref: '#/definitions/badge'
        - name: badgeId
          in: path
          description: Id of the badge that needs to be updated
          required: true
          type: integer
          format: int64
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: badge response
          schema:
            $ref: '#/definitions/badge'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    delete:
      summary: Delete an existing badge
      description: |
        Delete an existing badge
      operationId: DeleteBadge
      produces:
        - application/json
      parameters:
        - name: badgeId
          in: path
          description: Id of the badge that needs to be deleted
          required: true
          type: integer
          format: int64
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '204':
          description: badge deleted successfully
        '404':
          description: Badge not found
        '400':
          description: Invalid Id supplied
...
  /events:
    post:
      summary: Post events on the platform
      description: |
        Creates a new event
      operationId: addEvent
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: event
          in: body
          description: event object to add to the platform
          required: true
          schema:
            $ref: '#/definitions/event'
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: event response
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /application:
    post:
      summary: Post new application on the platform
      description: |
        Creates a new application
      operationId: addApplication
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: application
          in: body
          description: application object to add to the platform
          required: true
          schema:
            $ref: '#/definitions/application'
      responses:
        '200':
          description: application response
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/application'
    get:
      summary: Returns a the application token
      description: |
        Returns the application token
      operationId: loginApplication
      produces:
        - application/json
      parameters:
        - name: application
          in: body
          description: application object to add to the store
          required: true
          schema:
            $ref: '#/definitions/application'
      responses:
        '200':
          description: pointScale response
          schema:
            $ref: '#/definitions/pointScale'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    delete:
      summary: Delete an existing application
      description: |
        Delete an existing application
      operationId: DeleteApplication
      produces:
        - application/json
      parameters:
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '204':
          description: application deleted successfully
        '404':
          description: application not found
        '400':
          description: Invalid Id supplied
...
  /leaderboard:
    get:
      summary: Returns the leaderboard
      description: |
        Returns the leaderboard
      operationId: leaderboard
      produces:
        - application/json
      parameters:
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: leaderboard response
          schema:
            type: array
            items:
              $ref: '#/definitions/user'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
...
definitions:
  badge:
    type: object
    required:
      - id
      - name
      - app
      - userId
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      app:
        type: integer
        format: int64
      tag:
        type: string
      userId:
        type: integer
        format: int64
  pointScale:
    type: object
    required:
      - id 
      - name
      - app
      - userId
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      app:
        type: integer
        format: int64
      description:
        type: string
      token:
        type: string
      userId:
        type: integer
        format: int64
  user:
    type: object
    required:
      - id
      - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
  event:
    type: object
    required:
      - id 
      - userId
      - pointScale
      - increase
    properties:
      id:
        type: integer
        format: int64
      userId:
        type: integer
        format: int64
      pointScale:
        type: string
      increase:
        type: integer
        format: int64
  application:
    type: object
    required:
      - id 
      - name
      - password
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      password:
        type: string
  errorModel:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string